+++ lib/src/db/model_db_impl.dart
@@ -366,8 +366,17 @@ class ModelDBImpl implements ModelDB {
     return properties;
   }
 
-  bool _isExpandoClass(mirrors.ClassMirror modelClass) =>
-      modelClass.isSubtypeOf(mirrors.reflectClass(ExpandoModel));
+  bool _isExpandoClass(mirrors.ClassMirror modelClass) {
+    while (modelClass.superclass != modelClass) {
+      if (modelClass.reflectedType == ExpandoModel) {
+        return true;
+      } else if (modelClass.reflectedType == Model) {
+        return false;
+      }
+      modelClass = modelClass.superclass;
+    }
+    throw StateError('This should be unreachable.');
+  }
 }
 
 class _ModelDescription<T extends Model> {